+++ src/algo/mod.rs
@@ -8,6 +8,7 @@ pub mod astar;
 pub mod bellman_ford;
 pub mod dijkstra;
 pub mod dominators;
+pub mod edmonds_karp;
 pub mod feedback_arc_set;
 pub mod floyd_warshall;
 pub mod isomorphism;
@@ -36,6 +37,7 @@ use crate::visit::{Data, IntoNodeReferences, NodeRef};
 pub use astar::astar;
 pub use bellman_ford::{bellman_ford, find_negative_cycle};
 pub use dijkstra::dijkstra;
+pub use edmonds_karp::edmonds_karp;
 pub use feedback_arc_set::greedy_feedback_arc_set;
 pub use floyd_warshall::floyd_warshall;
 pub use isomorphism::{
